+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1 - 21/08/2022
+## [0.0.1]
 
-* Context,String Extension first version.
+- Context,String Extension first version.
+
+## [0.0.3]
+- Padding Extension